requirements.txt
google-genai
python-dotenv
PyPDF2
youtube-transcript-api
requests
